Integra a função de atulizar Funcionario a interface grafica, capturando os dados da tabela Listar funcionario.

+++ src/com/app/budega/controller/ListarFuncionarioController.java
@@ -2,19 +2,23 @@ package com.app.budega.controller;
 
 import com.app.budega.App.Main;
 import com.app.budega.dao.FuncionarioDAO;
+import com.app.budega.App.AtualizarFuncionarioMain;
 import com.app.budega.model.Funcionario;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
-import org.w3c.dom.ls.LSOutput;
+import javafx.stage.Stage;
 
 import java.net.URL;
 import java.sql.SQLException;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 public class ListarFuncionarioController implements Initializable {
@@ -58,8 +62,21 @@ public class ListarFuncionarioController implements Initializable {
         Funcionario funcionario = tabelaFuncionario.getSelectionModel().getSelectedItem();
         String cpf = funcionario.getCpf();
         try {
-            funcionarioDAO.deleteFuncionario(cpf);
-            initTable();
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Deletar Funcionarios");
+            alert.setHeaderText("Deseja realmente excluir ?");
+            alert.setContentText(funcionario.getNome());
+
+            Optional<ButtonType> resultado = alert.showAndWait();
+            if (resultado.get() == ButtonType.OK){
+                funcionarioDAO.deleteFuncionario(cpf);
+                initTable();
+            }else {
+                Alert alert1 = new Alert(Alert.AlertType.INFORMATION);
+                alert1.setTitle("Deletar Funcionarios");
+                alert1.setHeaderText("Nenhum funcionario selecionado.");
+                alert1.setContentText("Slecione um selecionado.");
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         } catch (ClassNotFoundException e) {
@@ -69,7 +86,15 @@ public class ListarFuncionarioController implements Initializable {
 
     @FXML
     void atualizarFuncionario(ActionEvent event){
-        Main.mudarTelas("atualizarFuncionario");
+        Funcionario funcionario = tabelaFuncionario.getSelectionModel().getSelectedItem();
+        AtualizarFuncionarioMain atualizarFuncionarioMain = new AtualizarFuncionarioMain(funcionario);
+        try {
+            atualizarFuncionarioMain.start(new Stage());
+            Main main = new Main();
+        }catch (Exception e){
+
+        }
+
     }
 }
